requirements.txt
python-telegram-bot==20.7
PyYAML==6.0.1
google-generativeai==0.3.2